A whole bag of feature improvements
- Automatic derivation for ConfDecoder
- Automatic derivation of Surface[T], to auto-generate docs and --help

I debated a lot with myself whether I should sunset this project or not.
On one hand, I would love to minimize my maintenance burden by building
on top of existing libraries. It would make a lot of sense to stop
metaconfig altogether and use something like Circe. On the other hand,
I feel like I would end up doing as much work building custom
configuration features on top of circe. The default circe error messages
are not so helpful, and I would need some way to automatically fail on
typos.

I decided to first experiment with implementing these macros on top of
metaconfig.Conf and see how it goes. This comit does not eliminate the
potential to switch to circe later down the road and the macros in this
repo can easily be tweaked when that happens. For now, this commit
solves two urgent needs in scalafix without breaking any existing code

* writing configuration for scalafix rules is plain tedious
* keeping the docs on configuration options up-to-date on the website
  is hard, causing a bad user experience. With the new
  metadoc.Surface[T], it should be possible to programmatically generate
  docs from the annotations on case fields.

I would love to experiement with a metaconfig-cli helper to convert command
line arguments to metaconfig.Conf. This would make it possible to drop
the dependency on case-app (and shapeless transitively).

A lot of work in this PR is inspired by code from other repos, including

- circe
- airframe
- case-app
- monix