Remove metaconfig-hocon
metaconfig-hocon is an experimental module that uses a custom HOCON parser
based on fastparse. However, it's complicated to maintain
fastparse-based projects for 2.11, 2.12, and 2.13, since fastparse
dropped 2.11 when it added 2.13 and the API is different.

Given that sconfig provides a good cross-platform solution, let's use
that instead of maintaining our own parser.