Add support to mix positional arguments with flags.
Previously, it was necessary to pass in all flags before the positional
arguments. The arguments `command positional_argument --flag` would be
parsed as two positional arguments "positional_argument" and "--flag".
Now, flags can be mixed with positionals so that "--flag" becomes parsed
as a flag instead of a positional argument.