Automatically support `--no` prefix for boolean flags.
Previously, if you had boolean flag like `sources: Boolean` with default
value true then there was no way for cli users to set the value to
false because `--sources` can only set the boolean to true. Now, users
can pass the flag `--no-sources` to make the boolean value false.

This diff ended up being trickier to implement than I expected because
we need to make sure that it's still OK to name fields with a `--no-`
prefix, such as `noSources: Boolean`. This feature is implemented in a
fully backwards compatible way by always trying first to use the default
parsing logic and only falling back to guessing the addition or removal
of the `--no-` prefix. This means that a boolean flag named `noSources:
Boolean` can also be overriden with the `--sources` flag.