Rename to MOpt
Metaconfig started originally as a "configuration library" and today it
still is 90% a configuration library. However, I primarily use
metaconfig today as a command-line parsing library. This name change
is meant to reflect this change

* make it easier for users to discover MOpt
* make it easier to explain and document MOpt

Additionally, this rename makes it possible to introduce some breaking
changes that are long overdue in Metaconfig. For example, I want to
change the `ConfDecoder[T].read()` signature to accept a `ConfContext`
instead of `Conf` so that we can make it easier to implement certain
features like custom transformations on `Conf.Obj` key names.