Levenshtein: change how a candidate is selected
Previously, we selected by the minimum Levenshtein distance, then used
additional logic to compute a ratio (which may not monotonically track
the distance values). Instead, sort by the final ratio.

As the tests can now attest, the current computation of ratio might be
too optimistic, as suggestions are not as tight.